Stage and Turn Control
Implemented the scripts for the stage and turn control so that the game can easily keep track of which stage and whos turn it is. I also made the stage control script expandable so that the number of stages in the game can be increased or decreased according to the game design.